requirements.txt
pandas
python-dotenv
streamlit
langchain
openai
ollama
langchain-openai
langchain-ollama